Delete round service and tests
