--- packages/api/vitest.config.ts
@@ -1,11 +1,0 @@
-import tsconfigPaths from 'vite-tsconfig-paths';
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-  plugins: [tsconfigPaths()],
-  test: {
-    include: ['tests/**/*.{test,spec}.{ts,tsx}'],
-    globals: true,
-    environment: 'node',
-  },
-});
